pyproject.toml
[project]
name = "gridtokenx-forecase-ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.111.0",
    "h5py>=3.16.0",
    "lightgbm==4.3.0",
    "matplotlib==3.8.4",
    "mlflow>=3.11.1",
    "notebook==7.1.3",
    "numba>=0.65.1",
    "numpy==1.26.4",
    "openpyxl>=3.1.5",
    "optuna>=4.8.0",
    "pandapower>=3.1.2",
    "pandas==2.2.2",
    "pyarrow==16.0.0",
    "pypsa>=1.2.0",
    "pyyaml==6.0.1",
    "scikit-learn==1.4.2",
    "tables>=3.11.1",
    "torch==2.3.0",
    "tqdm>=4.67.3",
    "uvicorn==0.29.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "httpx>=0.27",
]
geo = [
    "sqlalchemy>=2.0",
    "geoalchemy2>=0.14",
    "psycopg2-binary>=2.9",
    "geopandas>=0.14",
]
